+++ javaScript/projects/weather-app/src/components/tabs/unitsystems.js
@@ -1,7 +1,22 @@
 import importAll from '../../helpers/importAll';
 
-export default {
+export const unitSystem = {
   icons: importAll(require.context('../../assets/icons', false, /\.svg$/)),
+  get(key) {
+    return this.unitSystem[key];
+  },
+  setIcon(key) {
+    return this.icons.files[Object.keys(this.icons.files).find((iconKey) => iconKey.includes(key))];
+  },
+  roundValue(value) {
+    return Math.round(value);
+  },
+  setValue(obj, ...args) {
+    return this.roundValue(obj[`${args[0]}${this.get(args[1])}`]);
+  },
+};
+
+const unitSystems = {
   metric: {
     temp: 'c',
     speed: 'kph',
@@ -16,16 +31,8 @@ export default {
     pressure: 'in',
     distance: 'miles',
   },
-  get(key) {
-    return this.unitSystem[key];
-  },
-  setIcon(key) {
-    return this.icons.files[Object.keys(this.icons.files).find((iconKey) => iconKey.includes(key))];
-  },
-  roundValue(value) {
-    return Math.round(value);
-  },
-  setValue(obj, ...args) {
-    return this.roundValue(obj[`${args[0]}${this.get(args[1])}`]);
-  },
 };
+
+export function setUnitSystem(selection) {
+  Object.assign(unitSystem, { unitSystem: unitSystems[selection] });
+}